% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ui_chooseInput.R
\name{choose_input}
\alias{choose_input}
\title{Create radio buttons of the run mode in main Run KVFinder page}
\usage{
choose_input()
}
\description{
Create radio buttons of the run mode in main Run KVFinder page
}
