% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pdbProcess_.R
\name{get_ligand_pdb}
\alias{get_ligand_pdb}
\title{Get only target ligand PDB}
\usage{
get_ligand_pdb(pdb_input, ligand_name)
}
\arguments{
\item{pdb_input}{Pdb path}

\item{ligand_name}{Name of the ligand}
}
\description{
Separate from protein/nucleic PDB the target ligand
}
